Now compiles with MSVC

+++ src/popups/EmuFileExporterPopup.cpp
@@ -95,7 +95,7 @@ bool ebox::EmuFileExporterPopup::customDraw()
     m_tempoLabel.process();
     m_tempo.process();
 
-    ImGui::BeginChild("file_exporter_custom_length", {-1, 60}, true, 0);
+    ImGui::BeginChild("file_exporter_custom_length", {-1.f, 60.f}, true, 0);
     m_useCustomLength.process();
     if(m_useCustomLength.isChecked())
     {
@@ -108,7 +108,7 @@ bool ebox::EmuFileExporterPopup::customDraw()
     }
     ImGui::EndChild();
 
-    ImGui::BeginChild("file_exporter_voices", {-1, 40 + (m_file.getVoices()->size() * 30)}, true, 0);
+    ImGui::BeginChild("file_exporter_voices", {-1.f, (float)(40 + (m_file.getVoices()->size() * 30))}, true, 0);
     ImGui::Text("Voices:");
     for(auto &voice : *m_file.getVoices())
     {
